requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
filelock==3.20.0
fsspec==2025.10.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
Pygments==2.19.2
regex==2025.10.23
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
triton==3.5.0
typing_extensions==4.15.0
urllib3==2.5.0
